Bug Hunting
Added a code check for build menu labels so they are accurate.

Made the gold cost a negative to showcase loss of it.

Changed the damage icon to a sword to indicate damage better.

Added all towers to new system, healer towers won't spawn. Crashes for some reason not sure yet.