Currency and Code Tweaks
I added an global script called Commands. I figured it'd be a good way to hold some commands we might end up using a bunch for easier access.

Commands currently has a get_root function that just grabs the root of the scene.

I changed the currency to gold and mana so it's clear which is which.

Updated Icons for Gold and Currency as well as added the mana icon. I guess I forgot to push those last time which is on me.