Changed How Towers are Created
This should clear up some minor bugs, but also make it a lot easier to keep track of the towers and their stats.

Next step would be to remove the "Towers" node completely, as they just house data that could be stored a different way. Possibly in a "Game Dictionary" type .gd file? Idk.